Refactored services into a service director at root level of project
